Zvk: Temporarily disable the P extension
Zvkns, Zvknh[ab], Zvkg, Zvksed and Zvksh instructions utilize the 0x77
opcode assigned to the P extension. However there is no point in having
both P and Zvk enabled, those extensions are meant to be exclusive.

Currently Spike does not appear to have support for having 2 extensions
with aliasing encodings. The search algorithm is built so that the first
entry which matches the bit match+mask is chosen, without checking the
list of runtime-enabled extensions.

At some point this should be fixed in Spike but for the purpose of
extension ratification we disable the P extension to work around
this limitation.

Signed-off-by: Eric Gouriou <ego@rivosinc.com>
Co-authored-by: Stanislaw Kardach <kda@semihalf.com>
Signed-off-by: Eric Gouriou <ego@rivosinc.com>